feat: 거리 구하는 로직 추가
거리 구하는 로직 임시로 추가, 일부 자료형 변경

+++ Backend/ssaeng-go-jip/src/main/java/com/ssaenggojip/property/entity/Property.java
@@ -28,11 +28,11 @@ public class Property {
     private PropertyType propertyType;
 
     @Column(nullable = false)
-    private Integer price;
+    private Long price;
 
-    private Integer rentPrice;
+    private Long rentPrice;
 
-    private Integer maintenancePrice;
+    private Long maintenancePrice;
 
     @Column(nullable = false, precision = 10, scale = 2)
     private Double exclusiveArea;
